# docs/swagger.yaml
basePath: /
definitions:
  handler.ChangePasswordRequest:
    properties:
      email:
        example: shuser@maildrop.cc
        type: string
      new_password:
        example: Pa$$word1
        maxLength: 20
        minLength: 8
        type: string
      password:
        example: Pa$$w0rd!
        maxLength: 20
        minLength: 8
        type: string
    required:
    - email
    - new_password
    - password
    type: object
  handler.LoginRequest:
    properties:
      email:
        example: shuser@maildrop.cc
        type: string
      password:
        example: Pa$$w0rd!
        type: string
    required:
    - email
    - password
    type: object
  handler.RegisterRequest:
    properties:
      email:
        example: shuser@maildrop.cc
        type: string
      password:
        example: Pa$$w0rd!
        maxLength: 20
        minLength: 8
        type: string
      username:
        example: user1234
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is account service for ServiceHub
  title: ServiceHub Account API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      parameters:
      - description: Login Request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      responses: {}
      summary: Login account
  /api/auth/logout:
    post:
      responses: {}
      summary: Logout account
  /api/auth/profile:
    get:
      responses: {}
      summary: Profile of account
  /api/auth/register:
    post:
      parameters:
      - description: Register Request
        in: body
        name: RegisterRequest
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      responses: {}
      summary: Register an account
  /api/auth/reset-password:
    post:
      parameters:
      - description: Reset Password Request
        in: body
        name: resetPasswordRequest
        required: true
        schema:
          $ref: '#/definitions/handler.ChangePasswordRequest'
      responses: {}
      summary: Reset password
produces:
- application/json
schemes:
- http
swagger: "2.0"